+++ x/concentrated-liquidity/tick_test.go
@@ -18,13 +18,27 @@ import (
 
 const validPoolId = 1
 
+var (
+	defaultTickInfo = model.TickInfo{
+		LiquidityGross: DefaultLiquidityAmt,
+		LiquidityNet:   DefaultLiquidityAmt,
+		SpreadRewardGrowthOppositeDirectionOfLastTraversal: DefaultSpreadRewardAccumCoins,
+		UptimeTrackers: model.UptimeTrackers{List: wrapUptimeTrackers(getExpectedUptimes().hundredTokensMultiDenom)},
+	}
+
+	defaultTick = genesis.FullTick{
+		TickIndex: defaultTickIndex,
+		Info:      defaultTickInfo,
+	}
+
+	defaultTickWithoutPoolId = genesis.FullTick{
+		TickIndex: defaultTickIndex,
+		Info:      defaultTickInfo,
+	}
+)
+
 func withTickIndex(tick genesis.FullTick, tickIndex int64) genesis.FullTick {
 	tick.TickIndex = tickIndex
-	return tick
-}
-
-func withPoolId(tick genesis.FullTick, poolId uint64) genesis.FullTick {
-	tick.PoolId = poolId
 	return tick
 }
 
@@ -662,8 +676,6 @@ func (s *KeeperTestSuite) TestCrossTick() {
 }
 
 func (s *KeeperTestSuite) TestGetTickLiquidityForFullRange() {
-	defaultTick := withPoolId(defaultTick, defaultPoolId)
-
 	tests := []struct {
 		name        string
 		presetTicks []genesis.FullTick
@@ -803,9 +815,9 @@ func (s *KeeperTestSuite) TestGetTickLiquidityForFullRange() {
 			s.SetupTest()
 
 			// Create a default CL pool
-			s.PrepareConcentratedPool()
+			pool := s.PrepareConcentratedPool()
 			for _, tick := range test.presetTicks {
-				s.App.ConcentratedLiquidityKeeper.SetTickInfo(s.Ctx, tick.PoolId, tick.TickIndex, &tick.Info)
+				s.App.ConcentratedLiquidityKeeper.SetTickInfo(s.Ctx, pool.GetId(), tick.TickIndex, &tick.Info)
 			}
 
 			liquidityForRange, err := s.App.ConcentratedLiquidityKeeper.GetTickLiquidityForFullRange(s.Ctx, defaultPoolId)
@@ -816,8 +828,6 @@ func (s *KeeperTestSuite) TestGetTickLiquidityForFullRange() {
 }
 
 func (s *KeeperTestSuite) TestGetTickLiquidityNetInDirection() {
-	defaultTick := withPoolId(defaultTick, defaultPoolId)
-
 	tests := []struct {
 		name        string
 		presetTicks []genesis.FullTick
@@ -1200,7 +1210,7 @@ func (s *KeeperTestSuite) TestGetTickLiquidityNetInDirection() {
 			// Create a default CL pool
 			pool := s.PrepareConcentratedPool()
 			for _, tick := range test.presetTicks {
-				s.App.ConcentratedLiquidityKeeper.SetTickInfo(s.Ctx, tick.PoolId, tick.TickIndex, &tick.Info)
+				s.App.ConcentratedLiquidityKeeper.SetTickInfo(s.Ctx, test.poolId, tick.TickIndex, &tick.Info)
 			}
 
 			// Force initialize current sqrt price to 1.
@@ -1335,105 +1345,6 @@ func (s *KeeperTestSuite) TestValidateTickRangeIsValid() {
 				s.Require().ErrorContains(err, test.expectedError.Error())
 			} else {
 				s.Require().NoError(err)
-			}
-		})
-	}
-}
-
-func (s *KeeperTestSuite) TestGetAllInitializedTicksForPool() {
-	const (
-		// chosen randomly
-		defaultPoolId = 676
-	)
-
-	defaultTick := withPoolId(defaultTick, defaultPoolId)
-
-	tests := []struct {
-		name                   string
-		preSetTicks            []genesis.FullTick
-		expectedTicksOverwrite []genesis.FullTick
-		expectedError          error
-	}{
-		{
-			name:        "one positive tick per pool",
-			preSetTicks: []genesis.FullTick{defaultTick},
-		},
-		{
-			name:        "one negative tick per pool",
-			preSetTicks: []genesis.FullTick{withTickIndex(defaultTick, -1)},
-		},
-		{
-			name:        "one zero tick per pool",
-			preSetTicks: []genesis.FullTick{withTickIndex(defaultTick, 0)},
-		},
-		{
-			name: "multiple ticks per pool",
-			preSetTicks: []genesis.FullTick{
-				defaultTick,
-				withTickIndex(defaultTick, -1),
-				withTickIndex(defaultTick, 0),
-				withTickIndex(defaultTick, -200),
-				withTickIndex(defaultTick, 1000),
-				withTickIndex(defaultTick, -999),
-			},
-			expectedTicksOverwrite: []genesis.FullTick{
-				withTickIndex(defaultTick, -999),
-				withTickIndex(defaultTick, -200),
-				withTickIndex(defaultTick, -1),
-				withTickIndex(defaultTick, 0),
-				defaultTick,
-				withTickIndex(defaultTick, 1000),
-			},
-		},
-		{
-			name: "multiple ticks per multiple pools",
-			preSetTicks: []genesis.FullTick{
-				defaultTick,
-				withTickIndex(defaultTick, -1),
-				withPoolId(withTickIndex(defaultTick, 0), 3),
-				withTickIndex(defaultTick, -200),
-				withTickIndex(defaultTick, 1000),
-				withTickIndex(defaultTick, -999),
-				withPoolId(withTickIndex(defaultTick, -4), 90),
-				withTickIndex(defaultTick, 33),
-				withPoolId(withTickIndex(defaultTick, 44), 1200),
-				withPoolId(withTickIndex(defaultTick, -1000), 3),
-				withTickIndex(defaultTick, -1234),
-				withPoolId(withTickIndex(defaultTick, 1000), 3), // duplicate for another pool.
-			},
-			expectedTicksOverwrite: []genesis.FullTick{
-				withTickIndex(defaultTick, -1234),
-				withTickIndex(defaultTick, -999),
-				withTickIndex(defaultTick, -200),
-				withTickIndex(defaultTick, -1),
-				defaultTick,
-				withTickIndex(defaultTick, 33),
-				withTickIndex(defaultTick, 1000),
-			},
-		},
-	}
-
-	for _, test := range tests {
-		s.Run(test.name, func() {
-			s.SetupTest()
-
-			for _, tick := range test.preSetTicks {
-				s.App.ConcentratedLiquidityKeeper.SetTickInfo(s.Ctx, tick.PoolId, tick.TickIndex, &tick.Info)
-			}
-
-			// If overwrite is not specified, we expect the pre-set ticks to be returned.
-			expectedTicks := test.preSetTicks
-			if len(test.expectedTicksOverwrite) > 0 {
-				expectedTicks = test.expectedTicksOverwrite
-			}
-
-			// System Under Test
-			ticks, err := s.App.ConcentratedLiquidityKeeper.GetAllInitializedTicksForPool(s.Ctx, defaultPoolId)
-			s.Require().NoError(err)
-
-			s.Require().Equal(len(expectedTicks), len(ticks))
-			for i, expectedTick := range expectedTicks {
-				s.Require().Equal(expectedTick, ticks[i], "expected tick %d to be %v, got %v", i, expectedTick, ticks[i])
 			}
 		})
 	}
